Split test_soundex, testx_soundex
