+++ tests/test_soundex.py
@@ -11,7 +11,6 @@ from pythainlp.soundex import (
     soundex,
     udom83,
 )
-from pythainlp.soundex.sound import audio_vector, word_approximation
 
 
 class SoundexTestCase(unittest.TestCase):
@@ -82,9 +81,3 @@ class SoundexTestCase(unittest.TestCase):
         self.assertIsNotNone(prayut_and_somchaip("ณาญ"))
         self.assertIsNotNone(prayut_and_somchaip("กาง"))
         self.assertIsNotNone(prayut_and_somchaip("ว้าว"))
-
-    def test_word_approximation(self):
-        self.assertIsNotNone(word_approximation("รถ", ["รส", "รด", "คน"]))
-
-    def test_audio_vector(self):
-        self.assertIsNotNone(audio_vector("คน"))
